Add Oscillator header
